Re #27450 cherry-pick commit to prevent framework code using qt module
