+++ qt/python/mantidqt/widgets/plotconfigdialog/curvestabwidget/presenter.py
@@ -10,13 +10,13 @@ from __future__ import (absolute_import, unicode_literals)
 
 from matplotlib.collections import PolyCollection
 
+from mantid.plots.legend import LegendProperties
 from mantid.plots import Line2D, MantidAxes
 from mantidqt.utils.qt import block_signals
 from mantidqt.widgets.plotconfigdialog import get_axes_names_dict, curve_in_ax
 from mantidqt.widgets.plotconfigdialog.curvestabwidget import (
     CurveProperties, curve_has_errors, remove_curve_from_ax)
 from mantidqt.widgets.plotconfigdialog.curvestabwidget.view import CurvesTabWidgetView
-from mantidqt.widgets.plotconfigdialog.legendtabwidget import LegendProperties
 from workbench.plotting.figureerrorsmanager import FigureErrorsManager
